fix endscreen and add possibility to remove login-status to use joker with backspace or delete key
