+++ QuizPresentator/QuizPresentator.csproj
@@ -126,6 +126,7 @@
     <Compile Include="QuestionComponentBox.cs" />
     <Compile Include="ResultBox.cs" />
     <Compile Include="ResultBoxes.cs" />
+    <Compile Include="EmptyClass.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="FSharp.Core, Version=4.4.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a">
